jupyter/requirements.txt
jupyterlab>=3.2.5
numpy-quaternion>=2021.11.4.15.26.3
numpy>=1.21.3
torch>=1.10.0+cu113
torchvision>=0.11.1+cu113
Pillow>=8.4.0
matplotlib>=3.4.3
moviepy>=1.0.3